简单的策略测试

+++ qt-core/src/main/java/com/qklx/qt/core/response/Kline.java
@@ -5,21 +5,20 @@ import lombok.Data;
 import java.math.BigDecimal;
 
 /**
- * @Author ISME
+ * @Author yang
  * @Date 2018/1/14
  * @Time 11:35
+ * kline
  */
 @Data
 public class Kline {
-
-
     private long id;
-    private BigDecimal amount;
+    private Double amount;
     private int count;
-    private BigDecimal open;
-    private BigDecimal close;
-    private BigDecimal low;
-    private BigDecimal high;
-    private BigDecimal vol;
+    private Double open;
+    private Double close;
+    private Double low;
+    private Double high;
+    private Double vol;
 
 }
